(refactor): split timeline data into modular registry
- Move shared timeline types, category metadata, and filtering into a shared module
- Extract software-engineering and smartphone timelines into dedicated data files
- Add a typed timeline registry with exported keys for content schema validation
- Update timeline layout/components/tests to import from new timeline modules
- Add timeline authoring documentation and link it from course-materials authoring guide
