+++ src/layouts/materials/TimelineLayout.astro
@@ -2,7 +2,7 @@
 import type { CollectionEntry } from "astro:content";
 import Container from "../../components/ui/Container";
 import TimelineExplorer from "../../components/timeline/TimelineExplorer";
-import { getTimelineByKey } from "../../data/se-timeline";
+import { getTimelineByKey } from "../../data/timelines";
 
 interface Props {
   entry: CollectionEntry<"course-materials">;
